Implemented navigation in the MapsActivity
Also added comments. Only thing that still needs to be fixed is an error
that you get when you do not move the initial current location marker.
Nullpointerexception occurs. Idea: save original location and retrieve
the current marker location of the current user location. If equal,
break and tell the user to move the marker.